requirements.txt
ffmpeg
pyautogui
selenium
pillow 
easydev
matplotlib==3.6.1
pandas==1.5.1
opencv-python
extcolors
colormap
easydev
streamlit
